# Swift optimizations