--- a/KsApi.xcodeproj/xcshareddata/xcschemes/KsApi-iOS.xcscheme
+++ b/KsApi.xcodeproj/xcshareddata/xcschemes/KsApi-iOS.xcscheme
@@ -3,8 +3,8 @@
    LastUpgradeVersion = "0820"
    version = "1.3">
    <BuildAction
-      parallelizeBuildables = "YES"
-      buildImplicitDependencies = "YES">
+      parallelizeBuildables = "NO"
+      buildImplicitDependencies = "NO">
       <BuildActionEntries>
          <BuildActionEntry
             buildForTesting = "YES"