--- a/circle.yml
+++ b/circle.yml
@@ -24,8 +24,8 @@ test:
       swiftlint lint --strict --reporter json |
       tee $CIRCLE_ARTIFACTS/swiftlint-report.json
     - bin/test iOS 9.3
-    - bin/test iOS 10.1
-    - bin/test tvOS 10.1
+    - bin/test iOS 10.2
+    - bin/test tvOS 10.0
 experimental:
   notify:
     branches: