While trying to archive the app I noticed that there was a compiler crash when linking KsApi :( That sent me down a path to not only fix that, but also get some swift compiler optimizations done so that i could better test what was happening.

Some rough numbers I took:

Before optimizations: 1min 22secs (debug build), ~15 mins (archive build)
After optimizations: 45 secs (debug build), 2mins 30 secs (archive build)

Pretty significant! There are two things I did:

* Turned on xcode warnings for functions that take a long time to compile. I set the threshold to 500ms cause we have too many that take over 100ms, but that allowed me to find quite a few poor performers, including one that was taking 26 seconds!
* I noticed that `MockService` was taking 5+ mins to compile and another 5+ mins for code gen for release, and that was one of the reasons an archive took 15+ mins. The fix was simply to exclude it from building for release by doing `#if DEBUG`! Compile times sped up so much that I was even able to re-enable WMO, which @stephencelis and I couldn't figure out why it was taking so long many months ago.